Rename ndarray mod to ndarray_support
This avoids naming conflicts when testing.
