+++ .github/workflows/rust.yml
@@ -19,6 +19,6 @@ jobs:
     - name: Build
       run: cargo build --verbose
     - name: Run tests
-      run: cargo test --tests --verbose
+      run: cargo test --tests --verbose --all-features
     - name: Run doctests
-      run: cargo test --doc --verbose
+      run: cargo test --doc --verbose --all-features
